Moved application validators to their own file
